+++ src/components/accession2/index.tsx
@@ -1,67 +1,2 @@
-import { TabContext, TabList, TabPanel } from '@mui/lab';
-import { Box, Tab, Typography } from '@mui/material';
-import { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
-import { Accession2, getAccession2 } from 'src/api/accessions2/accession';
-import strings from 'src/strings';
-import TfMain from '../common/TfMain';
-import DetailPanel from './DetailPanel';
-
-export default function Accession2View(): JSX.Element {
-  const { accessionId } = useParams<{ accessionId: string }>();
-  const [selectedTab, setSelectedTab] = useState('detail');
-  const [accession, setAccession] = useState<Accession2>();
-
-  useEffect(() => {
-    const populateAccession = async () => {
-      const response = await getAccession2(parseInt(accessionId, 10));
-      setAccession(response);
-    };
-
-    if (accessionId !== undefined) {
-      populateAccession();
-    } else {
-      setAccession(undefined);
-    }
-  }, [accessionId]);
-
-  const handleChange = (event: React.SyntheticEvent, newValue: string) => {
-    setSelectedTab(newValue);
-  };
-
-  const tabStyles = {
-    textTransform: 'capitalize',
-    '&.Mui-selected': {
-      color: '#3A4445',
-      fontWeight: 600,
-    },
-  };
-
-  return (
-    <TfMain>
-      <Box padding={3}>
-        <Typography>{accession?.accessionNumber}</Typography>
-        <Typography color='#343A40' fontSize='24px' fontStyle='italic' fontWeight={500}>
-          {accession?.speciesScientificName}
-        </Typography>
-        <Typography color='#708284'>{accession?.speciesCommonName}</Typography>
-      </Box>
-      <Box sx={{ width: '100%' }}>
-        <TabContext value={selectedTab}>
-          <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
-            <TabList onChange={handleChange}>
-              <Tab label={strings.DETAIL} value='detail' sx={tabStyles} />
-              <Tab label={strings.HISTORY} value='history' sx={tabStyles} />
-              <Tab label={strings.VIABILITY_TESTING} value='viabilityTesting' sx={tabStyles} />
-            </TabList>
-          </Box>
-          <TabPanel value='detail'>
-            <DetailPanel accession={accession} />
-          </TabPanel>
-          <TabPanel value='history'>{strings.HISTORY}</TabPanel>
-          <TabPanel value='viabilityTesting'>{strings.VIABILITY_TESTING}</TabPanel>
-        </TabContext>
-      </Box>
-    </TfMain>
-  );
-}
+import Accession2View from './Accession2View';
+export { Accession2View };
